Move `AddressParser` instantiation to general `RDFServiceProvider`
